Use size_t for all lengths
Previously, uint32_t and size_t were used interchangeably. These types are
interchangeable on Xtensa gcc (uint32_t == size_t == unsigned int), but not on most
other 32-bit gcc architectures (where uint32_t == unsigned long, size_t ==
unsigned int).

+++ lib/include/mqtt_msg.h
@@ -63,9 +63,9 @@ enum mqtt_message_type {
 
 typedef struct mqtt_message {
     uint8_t *data;
-    uint32_t length;
-    uint32_t fragmented_msg_total_length;       /*!< total len of fragmented messages (zero for all other messages) */
-    uint32_t fragmented_msg_data_offset;        /*!< data offset of fragmented messages (zero for all other messages) */
+    size_t length;
+    size_t fragmented_msg_total_length;       /*!< total len of fragmented messages (zero for all other messages) */
+    size_t fragmented_msg_data_offset;        /*!< data offset of fragmented messages (zero for all other messages) */
 } mqtt_message_t;
 
 typedef struct mqtt_connection {
@@ -73,7 +73,7 @@ typedef struct mqtt_connection {
 
     uint16_t message_id;
     uint8_t *buffer;
-    uint32_t buffer_length;
+    size_t buffer_length;
 
 } mqtt_connection_t;
 
@@ -122,13 +122,13 @@ static inline int mqtt_get_retain(const uint8_t *buffer)
     return (buffer[0] & 0x01);
 }
 
-void mqtt_msg_init(mqtt_connection_t *connection, uint8_t *buffer, uint32_t buffer_length);
-bool mqtt_header_complete(uint8_t *buffer, uint32_t buffer_length);
-uint32_t mqtt_get_total_length(const uint8_t *buffer, uint32_t length, int *fixed_size_len);
-char *mqtt_get_publish_topic(uint8_t *buffer, uint32_t *length);
-char *mqtt_get_publish_data(uint8_t *buffer, uint32_t *length);
-uint16_t mqtt_get_id(uint8_t *buffer, uint32_t length);
-int mqtt_has_valid_msg_hdr(uint8_t *buffer, uint32_t length);
+void mqtt_msg_init(mqtt_connection_t *connection, uint8_t *buffer, size_t buffer_length);
+bool mqtt_header_complete(uint8_t *buffer, size_t buffer_length);
+size_t mqtt_get_total_length(const uint8_t *buffer, size_t length, int *fixed_size_len);
+char *mqtt_get_publish_topic(uint8_t *buffer, size_t *length);
+char *mqtt_get_publish_data(uint8_t *buffer, size_t *length);
+uint16_t mqtt_get_id(uint8_t *buffer, size_t length);
+int mqtt_has_valid_msg_hdr(uint8_t *buffer, size_t length);
 
 mqtt_message_t *mqtt_msg_connect(mqtt_connection_t *connection, mqtt_connect_info_t *info);
 mqtt_message_t *mqtt_msg_publish(mqtt_connection_t *connection, const char *topic, const char *data, int data_length, int qos, int retain, uint16_t *message_id);
